modules/brand-perceptions/requirements.txt
# Reddit API
praw==7.7.1

# YouTube API
google-api-python-client==2.108.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2

# Data processing
pandas==2.1.3
pyyaml==6.0.1

# Environment management
python-dotenv==1.0.0

# Rate limiting
ratelimit==2.2.1

# JSON handling
orjson==3.9.10
